Balanced first location A lot new features

+++ core/src/com/darkhouse/gdefence/Level/Tower/Projectile.java
@@ -92,7 +92,7 @@ public class Projectile extends GameActor{
         }else {
             position.set(targetV);
 //            tower.hitTarget(target, isMainProjectile);
-            tower.hitTarget(target, ((int) (tower.getDmg(target, isMainProjectile) * dmgMultiplayer)));
+            tower.hitTarget(target, /*((int) (*/tower.getDmg(target, isMainProjectile) * dmgMultiplayer)/*))*/;
             afterHit();
             Map.projectiles.remove(this);
         }
